change return type on add data and change parameter on remove data on domain

+++ domain/src/main/java/com/android/domain/usecase/AddPurchasedProductUseCase.kt
@@ -6,7 +6,7 @@ import com.android.domain.repository.BasicEcommerceRepository
 
 class AddPurchasedProductUseCase(private val repository: BasicEcommerceRepository) {
 
-    suspend fun execute(product: ProductPromo): Int = repository.addPurchasedProduct(product)
+    suspend fun execute(product: ProductPromo): Long = repository.addPurchasedProduct(product)
 
 
 }
